refactor out 'use_mosaiq' block
